Update the validation program

+++ parallel_matrix_mul_fifo.cpp
@@ -17,10 +17,11 @@ int main()
   {
     for (size_t col = 0; col < matrix_size; col++)
     {
-      src1[row][col] = row + col;
-      src2[row][col] = row + col;
+      src1[row][col] = static_cast<int>(row) + static_cast<int>(col);
+      src2[row][col] = static_cast<int>(row) + static_cast<int>(col);
     }
   }
+
   auto start_time = std::chrono::high_resolution_clock::now();
   matrixBenchMark.parallel_computing_fifo(src1, src2, dst, block_size);
   auto end_time = std::chrono::high_resolution_clock::now();
